kernel: work: check handler when submit to queue
Assert that the handler of a work is not NULL when submitting
it to the queue. This allows early detection of the
code that is submitting a non-NULL work with NULL handler to
the work queue (where it happens), rather than right before the
work item get executed in the queue (when it happens).

(cherry picked from commit 0cd135920a883c89437e9a754dc2a88f4cb39ee2)

Original-Signed-off-by: Yong Cong Sin <ycsin@meta.com>
GitOrigin-RevId: 0cd135920a883c89437e9a754dc2a88f4cb39ee2
Change-Id: I8ad96dd71369a33b9a722c1a8ee971920b781d02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4916718
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>