llext: Fix a bug in section linking
When relocating section symbol addresses the value where the
relocation is to be written is an offset into the section to load.

Simply rewriting it with the section address is not enough, we need
to write the address of the section with the offset into it.

(cherry picked from commit c678f25bc1ac5d5aad8f18779e2a578246febb8c)

Original-Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: c678f25bc1ac5d5aad8f18779e2a578246febb8c
Change-Id: Ib99a20b9d3ca72ab12f29c3359ec31e040b214ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4940265
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>