cmake: fix git dependency for version.h creation
Fix a variable spelling mistake which caused a required dependency to
not be set.

Together with this fix, then remove an unneeded CMake cache update.

The `ZEPHYR_GIT_INDEX-NOTFOUND` value is not really useful in this
case as a not set var will also ensure that Zephyr is checked to be
a git repo or not.
And when detected as a git repo, then `ZEPHYR_GIT_INDEX` is set
to correct value.

(cherry picked from commit 12c1fe4d8374147a8d57ac7bc3e1497f41c32d18)

Original-Fixes: #63610
Original-Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
GitOrigin-RevId: 12c1fe4d8374147a8d57ac7bc3e1497f41c32d18
Change-Id: If1d444ef177b113a204d8b9c4b48a3aeb2ec80d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4920712
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>