fuel_gauge: Remove keep alphabetized comment
The fuel gauge header asks to alphabetize the properties in order to keep
them organized and easy to read. However, this is difficult to enforce
without adding an adhoc script for just this header. In addition, there
aren't so many properties such that it's difficult to parse unsorted. Since
the alphabetization is already not followed and not likely to be easily
followed in the future, remove this ask from the header.

Remove "keep properties alphabetized" from the header.

(cherry picked from commit 771770a772318e49611eacf6947a40ad5d433e67)

Original-Signed-off-by: Aaron Massey <aaronmassey@google.com>
GitOrigin-RevId: 771770a772318e49611eacf6947a40ad5d433e67
Change-Id: I2adea55f3ed167e69f9e0df34afead2533e56044
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4925665
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>