kernel: Fixes for z_early_rand_get
The early random get function was making many wrong assumptions
about random subsys and entropy drivers. First, it was assuming
that entropy_get_entropy() would be ISR safe, that is not right,
the driver has an ISR safe callback and if it is not implemented
or not working it is not ok using the other callback.
Second, the fallback to the random subsys is even more problematic
since they can use kernel services to protect internal states and be
thread-safe.

Another incorrect thing in this function was the guard around it.
It was needed by features like stack randomization and stack canaries,
and not when those conditions were match. Just remove it and in case
it is not needed the linker will take care of it.

The drawback of this change is that in the absence of an entropy
generator with support to be called from ISR the randomness is very
weak.

(cherry picked from commit 6c0fad2888255e661925ef9f49b8b776844d809a)

Original-Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: 6c0fad2888255e661925ef9f49b8b776844d809a
Change-Id: Iaf3e4399edc6ef6fe0fa18d259e196bd82226313
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4938802
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>