nrf_bsim: Only connect once to Bsim Phy
Each MCU initialization hooks will attempt to call
into the API to connect to the bsim phy and, if the
user requested it thru command line, delay its boot.

Doing this for all MCUs is best to ensure it is done
even if there is no image for the other MCU,
but results in the calls being done twice if there is
2 MCUs images, which results in the simulator API giving
a warning about it being likely an error in the app.

To avoid this problem, let's have this be called
only once.

(cherry picked from commit 413a349264cab91c9a2bb322c32968344ee663df)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 413a349264cab91c9a2bb322c32968344ee663df
Change-Id: I0aa49673514354d2139c2c4cdfdc5087eb54cd59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4971876
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>