drivers: eth_mcux: cache must be dcache
Currently the Kconfig for eth_mcux selects nocache if
HAS_MCUX_CACHE is set. But, for platforms that have a flexspi
cache but not a CPU cache, this is invalid, so clarify this
in the Kconfig definition.

(cherry picked from commit 7d5d4ed907fad929fe71489adeaaf666390ac72d)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 7d5d4ed907fad929fe71489adeaaf666390ac72d
Change-Id: Idc86a5a6ffe85f6d507b14f774e3c0119a6383ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4938826
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>