drivers: pwm: pwm_pca9685 set_pre_scale when not in restart mode
The pca9685 driver assumes the chip will be in "restart" mode
after putting it to sleep. This is not necessarily the case, which
can cause setting the prescaler to fail.
This fix allows the precaler to always be set and only restart the pwm's
when the chip was actually in restart mode after being put to sleep.

(cherry picked from commit 54c62232fcd9f132a4aa1c7aac835f0256483494)

Original-Signed-off-by: Jaap Versteegh <j.r.versteegh@gmail.com>
GitOrigin-RevId: 54c62232fcd9f132a4aa1c7aac835f0256483494
Change-Id: Ib4268d9b827f9aeb319a6a4513f47147b068fcfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4999928
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>