mgmt/osdp: Fix off-by-one in buf len checks
Initially, the command/reply ID byte was not part of the data length
macros. But later, when it was changed to include it, the buffer length
checks was not adjusted. Due to this, we were not using the last byte in
the buffer. Fix this issue by correcting the condition.

(cherry picked from commit 5b24a8ad72c5dc37f14a4469be8683b057ed6f62)

Original-Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>
GitOrigin-RevId: 5b24a8ad72c5dc37f14a4469be8683b057ed6f62
Change-Id: I1c342350f464e13bf75372a25723a9acb88eb913
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5057244
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>