drivers: sensor: bq274xx: Configure or confirm chemistry profile
Both the BQ27421 and BQ27427 have a few preset Chemistry profiles.
For the BQ27421 there exists three variants of the IC, and for the BQ27427,
it can be configured. The chemistry profile among other things includes the
taper voltage, which is used to detect charge termination.

This adds an optional `chemistry-id` config option to the driver. On the
BQ27421, it will confirm that the correct variant of the IC is mounted,
and on the BQ27427, it will configure it with the correct value.

Side note: The reference manual for the BQ27427
(https://www.ti.com/lit/ug/sluucd5/sluucd5.pdf) currently contains some
errors and inconsistencies regarding these registers. The table on page 7
appears to be correct.

(cherry picked from commit ec202d852f8dffb3787349088c93266505d83f8b)

Original-Signed-off-by: Tobias Pisani <topisani@hamsterpoison.com>
GitOrigin-RevId: ec202d852f8dffb3787349088c93266505d83f8b
Change-Id: Iea38cfaeb750f5bbff8c02e27c5e151fc9b42186
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5016723
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>