github: compliance: fix python cache keys
The compliance and coding guidelines workflows are using the same pip
cache key, but they install a different set of packages. This is
probably a copy paste error from the documentation build workflow, which
has been changed to use a hash, and is causing one of the two steps to
restore from a incorrect cache.

Fix this by using a hash of the workflow file itself as a key, as that's
where the python package list is defined.

(cherry picked from commit 639d1f803b3db48941ce7c27e55ad130db67b547)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: 639d1f803b3db48941ce7c27e55ad130db67b547
Change-Id: Ie7890718b612a5a18b0cc102edfe215dabb11cdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4993103
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>