drivers: spi: dw: cleanup instantiation macro
This cleans up the instantiation macro. DBG_COUNTER was also removed
as that appears to be unnecessary. This also allows for if it is a
serial target to be configured from the devicetree.

(cherry picked from commit 909da582c52e906e02634e5b2dc9c749c5302e12)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: 909da582c52e906e02634e5b2dc9c749c5302e12
Change-Id: Idb75221996c1f4b581da433264e4b9149bb7c7ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5220897
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>