drivers: regulator: npm1300: Add boot state detection
Added boot state detection, so that the reference count is
set correctly and enable/disable work as expected.

With this fix it is no longer necessary to set
regulator-boot-on when the regulator is enabled in hardware.

(cherry picked from commit ec527228613e2768caaddedaa5cdea04951d3a06)

Original-Signed-off-by: Andy Sinclair <andy.sinclair@nordicsemi.no>
GitOrigin-RevId: ec527228613e2768caaddedaa5cdea04951d3a06
Change-Id: Ib8e37ed999b1ef2fd101f00724ad1cf048aa222c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5012410
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>