tests: kernel/smp_boot_delay: make test more generic
This makes the test a bit more generic so it can run on
other SMP enabled platforms. Though, we don't need CI
running this on every SMP platforms as this is not
testing the mechanism of bringing up a CPU. That is
being tested on the other SMP test. Therefore, only
enable qemu_x86_64 to be testing in CI so that this
feature will still be tested on CI.

(cherry picked from commit ad87d2667c3105b7de690af65da6eb08e2496c5a)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: ad87d2667c3105b7de690af65da6eb08e2496c5a
Change-Id: Ie57a75f4ed58e23ddc5f906a7810c2ad0161d7ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5211254
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>