intc: nxp_s32: use instance-based DT macros
At present, many of the NXP S32 shim drivers do not make use of
devicetree instance-based macros because the NXP S32 HAL relies on an
index-based approach, requiring knowledge of the peripheral instance
index during both compilation and runtime, and this index might not
align with the devicetree instance index.

The proposed solution in this patch eliminates this limitation by
determining the peripheral instance index during compilation
through macrobatics.

Note that for some peripheral instances is needed to define the
HAL macros of the peripheral base address because there are gaps
in the instances or there are SoCs with a single instance.

(cherry picked from commit 237ec65ad365dd027552da3823da2fcc414b5138)

Original-Signed-off-by: Manuel Argüelles <manuel.arguelles@nxp.com>
GitOrigin-RevId: 237ec65ad365dd027552da3823da2fcc414b5138
Change-Id: I191f4b22d433317f1809833772c39c9ab2b239ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5010154
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>