acpi: shell: Fix specifying string precision
The ACPI table signature is not null terminated, so a precision needs to
be provided in format strings. There was an attempt to do this, but it
was done in an incorrect way, which resulted in garbage characters
getting printed.

(cherry picked from commit 6e8a1f5859c85e62c6325d3ecb1cf2c61a42a2bf)

Original-Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
GitOrigin-RevId: 6e8a1f5859c85e62c6325d3ecb1cf2c61a42a2bf
Change-Id: I9554f6c0e7253c019dadfc79772e8d6ce1a318b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5009507
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>