hostap: kconfig: let entry be a menuconfig
wpa_supplicant has a lot of entries, and it appears in the top-level menu.
Let it be a menuconfig, so that it gets its own menu page with
all the options, instead of expanding in the top-level menu when
selected.

(cherry picked from commit ef5dda3378da64b61d03f09f28151755fc78f05a)

Original-Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>
Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: ef5dda3378da64b61d03f09f28151755fc78f05a
Change-Id: I6aaaf0bd71ea9b7e6d165c215d36cffc288ca56a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5383257
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Keith Short <keithshort@chromium.org>