drivers: gpio: make reset of TLE9104 optional
In some hardware designs it might happen that the reset signal
for the TLE9104 is not used only for this purpose, but instead for
instance to reset other devices at the same time. For such a hardware
design it is then necessary to make the reset GPIO optional. The reset
will have to be triggered earlier on.

(cherry picked from commit f26da17723084652f45016b77550ff49c7c74186)

Original-Signed-off-by: Benedikt Schmidt <benedikt.schmidt@embedded-solutions.at>
GitOrigin-RevId: f26da17723084652f45016b77550ff49c7c74186
Change-Id: I41a85231acea411a9879e428e9e23b2fc73647e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5240571
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>