doc: gsg: macOS: Include instructions to add homebrew to the path
Multiple users have failed to read the output from the Homebrew
installation script, which instructs the user how to add it to the
path. Include the instructions in the guide.

Also add a step to include the Homebrew Python executable to the path,
which allows then for invocation of python and pip as well as python3
and pip3.

(cherry picked from commit 318836af23aa1607342a019c01580fd87442f32b)

Original-Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>
GitOrigin-RevId: 318836af23aa1607342a019c01580fd87442f32b
Change-Id: I64fa4439380088192d95cd32fbe6cb1a51c9f4ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5020932
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>