samples shell/fs docs: Enable for native_sim
Instead of always enabling fuse for native_posix,
let's give users instructions on how to enable it
while building.

In the documentation: replace the references to native_posix
with native_sim, improve the links, mention the flash
simulator, and provide a separate subsection for FUSE access.

Add a separate test for the sample with fuse access enabled
and be sure it is possible to run both this and the basic
one with native_sim (was filtered out before for native_posix)

(cherry picked from commit 65fd7d42467225c3c9df013fb710587a62bbf874)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 65fd7d42467225c3c9df013fb710587a62bbf874
Change-Id: I6eda789a9661989a2a0957487355b4e3ccb8ab13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5041370
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>