drivers: i3c: add additional variable for num xfer for i3c
With I3C, when a controller does a read whether, through a private transfer
or a CCC, the target is responsible for issuing the End of Data through the
T bit, but the way the API is currently implemented, there is no way to
communicate this back up. For example, if a controller tries to read 100
Bytes (where the controller is to do an abort after 100 (or 101 bytes))
from a Target and the Target gives the EoD at the 42nd Byte, then there is
no way to currently know where the EoD occurred and how far it can be read
up to in the provided buffers.

This implements a num_xfer for `i3c_msg` and `i3c_ccc_payload` which the
driver is responsible for writing the total number of bytes transferred.

(cherry picked from commit 78e961998aa7ea17560e934b08b8489b4a300ea8)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: 78e961998aa7ea17560e934b08b8489b4a300ea8
Change-Id: If08e0a2c77fff9f28d4d3dc2b9832ecf50693413
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5084407
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>